chore: move data-proxy into lib/server
